+++ Resource.Infrastructure/Query/Projections/ResourceUpcasters.cs
@@ -48,7 +48,7 @@ namespace Resource.Microservice.Projections
                 public Guid ResourceId { get; set; }
                 public int ResourceNo { get; set; }
                 public string ResourceName { get; set; }
-                public decimal ResourcePrice { get; set; }
+                public int ResourcePrice { get; set; }
             }            
         }
     }
